Update all team members on tab list, fixes #29
When the members of a team changes, it's not enough to update the
added/removed player, all team members need to be updated on the tab
list. This is because we need to update the display of who is and isn't
in the same team as the player who was added/removed.