Use dropItem instead of dropItemNaturally
There is basically no reason to use dropItemNaturally. The name is
misleading, "naturally" has no real meaning, it's just a legacy name as
far as I can tell. We don't want random offsets to the drop location,
because it might lead to spawning items inside blocks, casuing them to
glitch out (at least in some versions of the game).