Run build job automatically
The problem with manual builds is that links to the latest successful
artifacts might not always work, if the build job has not been triggered
for the latest successful pipeline.

Another related problem is that artifacts are only kept for the latest
successful pipeline, not for the latest successful job. So if the build
job has not been triggered for the latest successful pipeline, the
artifacts are not kept and will be deleted after expire_in.